将content模块抽离了除了，search模块在front中，完成ElasticSearch一键同步商品索引，以及ActiveMQ发送“添加”“删除”消息，Redis因为是在前台用到，后台就没管了。改配置是在本地的。发布到公共服务器的话是要重新复制一份代码然后改配置吗？

+++ whysu-mall-manager/whysu-mall-manager-pojo/src/main/java/top/whysu/manager/pojo/TbMember.java
@@ -1,7 +1,6 @@
 package top.whysu.manager.pojo;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
 import java.util.Date;
 
 public class TbMember implements Serializable {
@@ -31,7 +30,7 @@ public class TbMember implements Serializable {
 
     private Integer points;
 
-    private BigDecimal balance;
+    private Long balance;
 
     public Long getId() {
         return id;
@@ -137,11 +136,11 @@ public class TbMember implements Serializable {
         this.points = points;
     }
 
-    public BigDecimal getBalance() {
+    public Long getBalance() {
         return balance;
     }
 
-    public void setBalance(BigDecimal balance) {
+    public void setBalance(Long balance) {
         this.balance = balance;
     }
 }
